requirements.txt
Django==1.11
djangorestframework==3.7.7
gunicorn==19.7.1
numpy==1.14.0
paho-mqtt==1.3.1
pytz==2017.3
